+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## Unreleased
 
-* Add static code diagnostics `always-remove-listener` and `avoid-unnecessary-setstate`.
+* Add static code diagnostics `always-remove-listener`, `avoid-wrapping-in-padding` and `avoid-unnecessary-setstate`.
 * Improve static code diagnostics `avoid-returning-widgets` and `avoid-unused-parameters`.
 * Remove deprecated `Lines of Executable Code` metric, use `Source lines of Code` instead.
 * Changed the supported `analyzer` version to `^1.7.0`.